requirements.txt
attrs==20.3.0
autopep8==1.5.5
aws-cdk.assets==1.91.0
aws-cdk.aws-applicationautoscaling==1.91.0
aws-cdk.aws-autoscaling-common==1.91.0
aws-cdk.aws-cloudwatch==1.91.0
aws-cdk.aws-codeguruprofiler==1.91.0
aws-cdk.aws-ec2==1.91.0
aws-cdk.aws-ecr==1.91.0
aws-cdk.aws-ecr-assets==1.91.0
aws-cdk.aws-efs==1.91.0
aws-cdk.aws-events==1.91.0
aws-cdk.aws-iam==1.91.0
aws-cdk.aws-kms==1.91.0
aws-cdk.aws-lambda==1.91.0
aws-cdk.aws-lambda-python==1.91.0
aws-cdk.aws-logs==1.91.0
aws-cdk.aws-s3==1.91.0
aws-cdk.aws-s3-assets==1.91.0
aws-cdk.aws-sqs==1.91.0
aws-cdk.aws-ssm==1.91.0
aws-cdk.cloud-assembly-schema==1.91.0
aws-cdk.core==1.91.0
aws-cdk.cx-api==1.91.0
aws-cdk.region-info==1.91.0
cattrs==1.1.2
constructs==3.3.48
jsii==1.23.0
publication==0.0.3
pycodestyle==2.6.0
python-dateutil==2.8.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
